@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq.Expressions;
-using System.Threading.Tasks;
-using DotNetCoreWebSample.Web.Models;
+﻿using DotNetCoreWebSample.Web.Models;
 
 namespace DotNetCoreWebSample.Web.Repositories
 {
